collect quaternion methods in a single file
